pyproject.toml
[tool.poetry]
name = "repl_nix_langchain-chatbot"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.10"
colorama = "^0.4.6"
python-dotenv = "^1.0.0"
langchain = "0.0.154"
pinecone-io = "0.0.1"
PyMuPDF = "1.22.2"
openai = "0.27.6"
pinecone-client = "2.2.1"
tiktoken = "0.3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
